requirements.txt
openai
requests
google-generativeai
ebooklib
markdown2
litellm